melo_tts/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
analytics-python==1.4.post1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
backoff==1.10.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
boto3==1.34.96
botocore==1.34.96
bpemb==0.3.5
build==1.2.1
CacheControl==0.14.0
cached_path==1.6.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
cn2an==0.5.22
conllu==4.5.3
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
Distance==0.1.3
distlib==0.3.8
docopt==0.6.2
dulwich==0.21.7
eng-to-ipa==0.0.2
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
flair==0.12.2
fonttools==4.51.0
fsspec==2024.2.0
ftfy==6.2.0
fugashi==1.3.0
future==1.0.0
g2p-en==2.1.0
g2pkk==0.1.2
gdown==4.4.0
gensim==4.3.2
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gradio==1.0.0
gradio_client==0.16.0
grpcio==1.63.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
hyperopt==0.2.7
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.0.0
installer==0.7.0
jaconv==0.3.4
jamo==0.4.1
Janome==0.5.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
langdetect==1.0.9
langid==1.1.6
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.42.0
loguru==0.7.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.5
monotonic==1.6
more-itertools==10.2.0
mpld3==0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.59.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
orjson==3.10.2
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
pip-tools==7.4.1
pkginfo==1.10.0
plac==1.4.3
platformdirs==4.2.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
pooch==1.8.1
pptree==3.1
proces==0.1.7
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pykakasi==2.2.1
PyNaCl==1.5.0
pyparsing==3.1.2
pypinyin==0.50.0
pyproject_hooks==1.1.0
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.2
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.8.0
SecretStorage==3.3.3
segtok==1.5.11
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.4.24
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.27.4
triton==2.3.0
trove-classifiers==2024.4.10
txtsplit==1.0.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==1.26.18
uvicorn==0.29.0
virtualenv==20.26.1
wasabi==0.10.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
Wikipedia-API==0.6.0
wrapt==1.16.0
zipp==3.18.1
